requirements.txt
django-cleanup==5.2.0